Use Relative Path for IDE
This resolves an issue where the user may choose to provide a non-root route builder to Pioneer.

For example, a user may pass a RouteBuilder through which has a relative path of `/api` which (when the API path of "graphql" is added) would place the resolver at /api/graphql.

Currently, we would still be looking at /graphql which would fail. This patch means it will now look relative to the playground endpoint and correctly maps to /api/graphql in my testing.