Update Schema Documentation for Array Relationship
Lost a little bit of time today trying to understand the compiler and runtime errors thrown when using an array return type from a function.

In the example, `friends` is a function which resolves an array of users.

With the current documentation, this will throw a compiler error complaining about the instance keypath being provided.

When replaced with `User.friends`, it compiles but at runtime you get a crash due to the unsupported array type. By boxing the type using the underlying TypeReference generic, this runtime error goes and it works as expected.

This syntax is demonstrated in the upstream Star Wars example but the Pioneer documentation did not include this which may cause confusion to new users.